Add logging of client-sent control chunks.
This fixes #34 on the server log side, but not the browser client side.

Also refactor the protocol module so it's shared between front and back end.
